Conversation flow, memory and recall
I implemented if-else statements and foreach loops to cycle through what the user is interested in so that it can be recalled later in the conversation. I also used these elements to make sure the AI chatbot recites information in simpler terms if the user is confused about a specific topic creating a seamless conversational flow.